+++ src/Validator/In.php
@@ -45,7 +45,7 @@
      * @return bool
      */
     public function isValid($value) {
-      $this->flushErrors();
+      $this->errors = [];
 
       if (in_array($value, $this->values)) {
         return true;